Struct rush::prompt::Prompt

pub struct Prompt {
    // some fields omitted
}

Prompt Struct containing prompt and cwd for use on every new line in Rusty

Methods

impl Prompt

fn new() -> Prompt

Instantiates a new Prompt with default values that will be overwritten when the configuration is updated in the main file for execution

fn update_prompt(&mut self)

Update Prompt Calls method in rush::config to update the current prompt Only needs to be called if using cd or su at this point in time

fn get_user_p(&self) -> String

Get User P Returns prompt to be displayed on the command line

fn get_cwd(&self) -> String

Get CWD Returns the CWD for use in prompts

fn update_cwd(&mut self)

Update CWD Used to update the CWD if using CD

fn print(&self)